pyproject.toml
[tool.poetry]
name = "paperview"
version = "0.1.0"
description = ""
authors = ["anaka <alex.naka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
attrs = "^22.1.0"
pre-commit = "^2.20.0"
pytest = "^7.2.0"
boto3 = "^1.26.32"
bs4 = "^0.0.1"
Pillow = "^9.3.0"
ipykernel = "^6.19.4"
notebook = "^6.5.2"
jupyter = "^1.0.0"
pandas = "^1.5.2"
seaborn = "^0.12.1"
pdfplumber = "^0.7.6"
tqdm = "^4.64.1"
click = "^8.1.3"
modal-client = "^0.44.603"
fastapi = {extras = ["all"], version = "^0.88.0"}
pydantic = "^1.10.2"
websockets = "^10.4"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
